pyproject.toml
[project]
name = "deep-research-sdk"
version = "0.1.3"
description = "Python SDK for Open Deep Research"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic>=2.12.5",
    "aiohttp>=3.13.3",
    "duckduckgo-search>=8.1.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.31",
    "tokenizers>=0.21.1",
    "litellm>=1.81.5",
    "markitdown>=0.1.4",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
]


[dependency-groups]
dev = ["pip>=25.0.1", "pre-commit>=4.2.0", "pytest>=8.3.5", "ruff>=0.11.2"]
